requirements.txt
streamlit
fastapi
uvicorn
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
pytube3
sqlalchemy
psycopg2-binary
python-multipart